Extract common animation variants to commonAnimations.ts and remove duplicates
