+++ src/animations/benefitsAnimations.ts
@@ -1,34 +1,6 @@
-export const containerVariants = {
-  hidden: { 
-    opacity: 0 
-  },
-  visible: {
-    opacity: 1,
-    transition: { 
-      staggerChildren: 0.2 
-    }
-  }
-};
+import { containerVariants, headerVariants } from './commonAnimations';
 
-export const headerVariants = {
-  hidden: { 
-    opacity: 0, 
-    y: -20 
-  },
-  visible: { 
-    opacity: 1, 
-    y: 0 
-  }
-};
-
-export const contentVariants = {
-  hidden: { 
-    opacity: 0 
-  },
-  visible: { 
-    opacity: 1 
-  }
-};
+export { containerVariants, headerVariants };
 
 export const ctaVariants = {
   hidden: {
