+++ src/animations/progressAnimations.ts
@@ -1,25 +1,7 @@
-export const containerVariants = {
-  hidden: { 
-    opacity: 0 
-  },
-  visible: {
-    opacity: 1,
-    transition: { 
-      staggerChildren: 0.2 
-    }
-  }
-};
+import { containerVariants, headerVariants } from './commonAnimations';
 
-export const headerVariants = {
-  hidden: { 
-    opacity: 0, 
-    y: -20 
-  },
-  visible: { 
-    opacity: 1, 
-    y: 0 
-  }
-};
+export { containerVariants, headerVariants };
+
 
 export const progressBarVariants = {
   hidden: { 
@@ -34,6 +16,7 @@ export const progressBarVariants = {
     }
   })
 };
+
 
 export const itemVariants = {
   hidden: {
