+++ src/components/CourseDescription/FeatureCard/FeatureCard.component.tsx
@@ -1,5 +1,5 @@
 import { motion } from "framer-motion";
-import { featureVariants } from "../../../animations/descriptionAnimations";
+import { itemWithDelayVariants } from "../../../animations/commonAnimations";
 
 type FeatureCardProps = {
   title: string;
@@ -10,7 +10,7 @@ type FeatureCardProps = {
 const FeatureCard = ({ title, description, index }: FeatureCardProps) => (
   <motion.div
     custom={index}
-    variants={featureVariants}
+    variants={itemWithDelayVariants}
     className="flex flex-col items-center gap-4 p-6 bg-white rounded-xl shadow-lg hover:shadow-xl transition-shadow duration-300 max-w-sm"
   >
     <div className="w-16 h-16 bg-primary/10 rounded-full flex items-center justify-center">
