+++ src/components/CourseTarget/TargetCards/TargetCards.component.tsx
@@ -1,11 +1,11 @@
 import { motion } from "framer-motion";
-import { cardsContainerVariants } from "../../../animations/targetAnimations";
+import { gridVariants } from "../../../animations/commonAnimations";
 import { COURSE_TARGET } from "../../../data/CourseTarget.data";
 import TargetCard from "../TargetCard/TargetCard.component";
 
 const TargetCards = () => (
   <motion.section 
-    variants={cardsContainerVariants}
+    variants={gridVariants}
     className="flex w-fit h-fit items-center justify-center flex-wrap gap-16 mt-20"
   >
     {COURSE_TARGET.map((target, index) => (
